WebApp/requirements.txt
antlr-ast==0.8.1
antlr4-python3-runtime==4.7.2
Authlib==1.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.2
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
pycparser==2.21
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
winnowing==0.2.1
WTForms==3.0.1
zipp==3.12.0
